@@ -1,32 +1,41 @@
 ﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Content;
-using System.Collections.Generic;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace Seihou
 {
-	
-	class QuestionState : State
+    class QuestionState : State
 	{
 		private readonly double _score;
 		private readonly Settings.Difficulty _difficulty;
 
-		Vector2 buttonSize = new Vector2(400, 60);
-        readonly List<Control> controls;
+		Vector2 buttonSize = new(400, 60);
+		readonly FormHost host = new();
+		const int controlSpacing = 100;
 
 		public QuestionState(StateManager sm, ContentManager cm, SpriteBatch sb, GraphicsDeviceManager gdm,double score, Settings.Difficulty difficulty) : base(sm, cm,sb,gdm)
 		{
 			_difficulty = difficulty;
 			_score = score;
 
-			controls = new List<Control>
+			host.AddControl(new Button(sb, No)
 			{
-				new Button(new Vector2(Global.screenWidth/2,Global.screenHeight/2-80),buttonSize,sb,Yes,"Yes please", 0, Button.Align.center),
-				new Button(new Vector2(Global.screenWidth/2,Global.screenHeight/2+80),buttonSize,sb,No,"NO TAKE ME TO MENU", 1, Button.Align.center),
-			};
+				TabIndex = 0,
+				Position = new Vector2(Global.screenWidth / 2 - buttonSize.X / 2, controlSpacing * 2),
+				Size = buttonSize,
+				Text = "No"
+			});
+
+			host.AddControl(new Button(sb, Yes)
+			{
+				TabIndex = 1,
+				Position = new Vector2(Global.screenWidth / 2 - buttonSize.X / 2, controlSpacing * 3),
+				Size = buttonSize,
+				Text = "Yes"
+			});
 		}
 
-		private void Yes(object sender)
+		private void Yes()
 		{
 			var data = new HighscoreState.PlayData
 			{
@@ -37,7 +46,7 @@ namespace Seihou
 			sm.ChangeState(new HighscoreState(sm, cm, sb, gdm, data));
 		}
 
-		private void No(object sender)
+		private void No()
 		{
 			sm.ChangeState(new MenuState(sm, cm, sb, gdm));
 		}
@@ -45,16 +54,13 @@ namespace Seihou
 		public override void Draw(GameTime gt)
 		{
 			var txt = "Do you want to save your PATHETIC score?";
-			sb.DrawString(ResourceManager.fonts["DefaultFont"],txt,new Vector2(Global.screenWidth/2,100),Color.White,0,new Vector2(ResourceManager.fonts["DefaultFont"].MeasureString(txt).X/2,0),1,SpriteEffects.None,0);
-			foreach (var c in controls) c.Draw(gt);
+			sb.DrawString(ResourceManager.fonts["DefaultFont"],txt,new Vector2(Global.screenWidth/2,controlSpacing),Color.White,0,new Vector2(ResourceManager.fonts["DefaultFont"].MeasureString(txt).X/2,0),1,SpriteEffects.None,0);
+			host.Draw(gt);
 		}
 
 		public override void Update(GameTime gt)
 		{
-			Cursor.Moved();
-			Global.buttonCount = controls.Count;
-			Button.ButtonKeyControl(gt);
-			foreach (var c in controls) c.Update(gt);
+			host.Update(gt);
 		}
 	}
 	
