@@ -1,164 +1,185 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
+using MonoGame;
+using System;
 
 namespace Seihou
 {
-	delegate void ButtonCallBack(object sender);
-
-	class Button : Control
+    internal class Button : Control
 	{
-        public enum Align
-        {
-            left,
-            center,
-            right,
+		public event EventHandler OnReleased;
+		public event EventHandler OnSelect;
+
+		public bool Enabled { get; set; } = true;
+
+		public Color TextColor { get; set; } = Color.White;
+		public Color TextColorSelected { get; set; } = new Color(170,170,170);
+		public Color TextColorOnMouseDown { get; set; } = new Color(110,110,110);
+		public Color TextColorTabIndex { get; set; } = new Color(230,230,150);
+		public Color BackgroundColor { get; set; } = Color.Transparent;
+		public Color BackgroundColorSelected { get; set; } = Color.Transparent;
+		public Color BackgroundColorOnMouseDown { get; set; } = Color.Transparent;
+		public Color BackgroundColorTabIndex { get; set; } = Color.Transparent;
+
+		public TextAlign Align { get; set; } = TextAlign.Center;
+		public TextGravity Gravity { get; set; } = TextGravity.Center;
+
+		public Vector2 Position { get; set; } = new Vector2(0, 0);
+		public Vector2 Size { get; set; } = new Vector2(100, 50);
+		public string Text { get; set; } = string.Empty;
+		public string Font { get; set; } = "DefaultFont";
+
+		private bool isSelected = false;
+		private bool isPressed = false;
+		private int lastTabIndex = -1;
+		private Vector2 lastMousePosition;
+		private Vector2 pressedAt;
+
+		private readonly TextLabel Label = new();
+
+        public Button(SpriteBatch sb) : base(sb)
+		{
         }
 
-        public Color selectedTextColor = Color.White;
-        public Color background3D = Color.Transparent;
-        public Vector2 background3DOffset = new Vector2(4, 4);
-        public ButtonCallBack onClicked;
-		public ButtonCallBack onHover;
+        public Button(SpriteBatch sb, Action onReleased) : this(sb)
+        {
+			OnReleased += (_,_) => onReleased();
+        }
 
-		private static KeyboardState oldKeyboardState;
-
-		Vector2 pos, size;
-		public string text;
-		readonly string font;	
-        bool hovering = false;
-		readonly Align align;
-		readonly int index;
-		bool selected = false;
-		static bool pressed = false;
-
-		static public bool AwaitFireKeyUp { private get; set; }
-
-		public Button(Vector2 pos, Vector2 size, SpriteBatch sb, ButtonCallBack onClicked, string text, int index = 0, Align align = Align.left, string font = "DefaultFont") : base(sb)
+		public Button(SpriteBatch sb, Action onReleased, Color simpleColor) : this(sb)
 		{
-			textColor = new Color(100, 100, 100);
-            this.font = font;
-            this.align = align;
-            this.onClicked = onClicked;
-            this.pos = pos - size / 2;
-			this.size = size;
-			this.sb = sb;
-			this.text = text;
-			this.index = index;
+			OnReleased += (_, _) => onReleased();
+			SetSimpleColorText(simpleColor);
+		}
+
+		public void SetSimpleColorText(Color c)
+        {
+			TextColor = c;
+			TextColorSelected = Darken(c, 0.5f);
+			TextColorOnMouseDown  = Darken(c, 0.2f);
+		}
+
+		public void SetSimpleColorBackground(Color c)
+		{
+			BackgroundColor = c;
+			BackgroundColorSelected = Darken(c, 0.5f);
+			BackgroundColorTabIndex = Darken(c, 0.5f);
+			BackgroundColorOnMouseDown = Darken(c, 0.2f);
+		}
+
+		private static Color Darken(Color c, float m)
+        {
+			return new Color((byte)(c.R * m), (byte)(c.G * m), (byte)(c.B * m), c.A);
+		}
+
+		protected virtual Color GetCurrentBackgroundColor()
+        {
+			if (isPressed)
+				return BackgroundColorOnMouseDown;
+
+			if (lastTabIndex == TabIndex)
+				return BackgroundColorTabIndex;
+
+			if (isSelected)
+				return BackgroundColorSelected;
+
+
+			return BackgroundColor;
+        }
+
+		protected virtual Color GetCurrentTextColor()
+		{
+			if (isPressed)
+				return TextColorOnMouseDown;
+
+			if (lastTabIndex == TabIndex)
+				return TextColorTabIndex;
+
+			if (isSelected)
+				return TextColorSelected;
+
+			return TextColor;
 		}
 
 		public override void Draw(GameTime gt)
 		{
-            MonoGame.Primitives2D.FillRectangle(sb, pos + background3DOffset, size, background3D, 0);
-            MonoGame.Primitives2D.FillRectangle(sb, pos, size, background, 0);
+			if (!Enabled) return;
 
-			#if DRAWCOLBOX
-			MonoGame.Primitives2D.DrawRectangle(sb, pos, size, Color.Red);
-			#endif
+            var font = ResourceManager.fonts[Font];
 
-			Vector2 orgin = new Vector2(0,0);
-            Vector2 place = new Vector2(0, 0);
-            orgin.Y = ResourceManager.fonts[font].MeasureString(text).Y / 2;
-            place.Y = size.Y / 2;
+            Primitives2D.FillRectangle(sb, Position, Size, GetCurrentBackgroundColor());
 
-            //Current font
-            var cf = ResourceManager.fonts[font]; 
-
-            switch (align)
-            {
-                case Align.left:
-                    orgin.X = 0;
-                    place.X = 0;
-                    break;
-
-                case Align.center:
-                    orgin.X = cf.MeasureString(text).X / 2;
-                    place.X = size.X / 2;
-                    break;
-
-                case Align.right:
-                    orgin.X = cf.MeasureString(text).X;
-                    place.X = size.X;
-                    break;
-            }
-			if(Global.selectedButton == index)
-			{
-				selected = true;
-			}
-			else
-			{
-				selected = false;
-			}
-				
-            sb.DrawString(cf, text,pos + place, selected ? selectedTextColor : textColor,0,orgin,1,SpriteEffects.None,0);
+			Label.Font = font;
+			Label.TextString = Text;
+			Label.Size = Size;
+			Label.Color = GetCurrentTextColor();
+			Label.Position = Position;
+			Label.TextAlign = Align;
+			Label.TextGravity = Gravity;
+			Label.Draw(sb);
 		}
 
 		public override void Update(GameTime gt)
 		{
-            MouseState mouseState = Mouse.GetState();
-            hovering = MouseOnButton();
-            
-            if (hovering && !Global.keyMode)
+			if (!Enabled) return;
+
+			if (isSelected)
+			{
+				OnSelect?.Invoke(this, EventArgs.Empty);
+			}
+		}
+
+		public override void SetTabIndex(int index)
+        {
+			lastTabIndex = index;
+		}
+
+		public override void MouseMove(Vector2 mouse)
+        {
+			isSelected = Intersects(mouse);
+
+			if (!isSelected)
             {
-				Global.selectedButton = index;
-				onHover?.Invoke(this);
-				if (Cursor.IsMouseLeftPressed())
-                    onClicked?.Invoke(this);
+				isPressed = false;
             }
 
-			if (pressed && (Global.selectedButton == index))
+			lastMousePosition = mouse;
+		}
+
+		public override void Press()
+        {
+			if (Intersects(lastMousePosition))
 			{
-				onClicked?.Invoke(this);
+				isPressed = true;
+				pressedAt = lastMousePosition;
 			}
+		}
+
+        public override void Release()
+        {
+			if (Intersects(lastMousePosition) && Intersects(pressedAt))
+            {
+				OnReleased?.Invoke(this, EventArgs.Empty);
+			}
+
+			isPressed = false;
         }
 
-		private bool MouseOnButton()
-		{
-			MouseState mouseState = Mouse.GetState();
-            return (((mouseState.X > pos.X) && (mouseState.X < pos.X + size.X)) && ((mouseState.Y > pos.Y) && (mouseState.Y < pos.Y + size.Y)));
+		private bool Intersects(Vector2 point)
+        {
+			return  
+				(point.X > Position.X) &&
+				(point.X < Position.X + Size.X) &&
+				(point.Y > Position.Y) &&
+				(point.Y < Position.Y + Size.Y);
 		}
 
-		public static void ButtonKeyControl(GameTime gameTime)
-		{
-			KeyboardState kb = Keyboard.GetState();
-
-			bool upWasDown = oldKeyboardState.IsKeyDown(Settings.GetKey("upKey"));
-			bool downWasDown = oldKeyboardState.IsKeyDown(Settings.GetKey("downKey"));
-			bool pressWasDown = oldKeyboardState.IsKeyDown(Settings.GetKey("shootKey"));
-
-			bool upUp = kb.IsKeyUp(Settings.GetKey("upKey"));
-			bool downUp = kb.IsKeyUp(Settings.GetKey("downKey"));
-			bool pressUp = kb.IsKeyUp(Settings.GetKey("shootKey"));
-
-			pressed = false;
-
-			if(downWasDown && downUp)
-			{
-				Global.selectedButton++;
-				Global.keyMode = true;
+        public override void OnKeyboardPress()
+        {
+            if (TabIndex == lastTabIndex)
+            {
+				OnReleased?.Invoke(this, EventArgs.Empty);
 			}
-			if (upWasDown && upUp)
-			{
-				Global.selectedButton--;
-				Global.keyMode = true;
-			}
-			if (pressWasDown && pressUp)
-			{
-				if(AwaitFireKeyUp)
-                {
-					AwaitFireKeyUp = false;
-                }
-				else
-                {
-					pressed = true;
-					Global.keyMode = true;
-				}
-			}
-
-			if (Global.selectedButton < 0) Global.selectedButton = 0;
-			if (Global.selectedButton >= Global.buttonCount) Global.selectedButton = Global.buttonCount -1;
-
-			oldKeyboardState = kb;
-		}
-	}
+        }
+    }
 }
