@@ -3,42 +3,46 @@ using Microsoft.Xna.Framework.Graphics;
 
 namespace Seihou
 {
-	class PickerButton : Control
-	{
-		int selected = 0;
-		readonly string[] answers;
-		public string question;
-		Button button;
-		readonly string text;
+    internal class PickerButton : Button
+    {
+        int selected = 0;
+        readonly string[] answers;
+        public string question;
+        readonly string text;
 
-		public PickerButton(Vector2 pos, Vector2 size, SpriteBatch sb, string text, string question, string startsWith, int index, params string[] answers) : base(sb)
-		{
-			for (int i = 0; i < answers.Length; ++i) selected = answers[i] == startsWith ? i : selected;
-			this.question = question;
-			this.answers = answers;
-			this.text = text;
-			button = new Button(pos, size, sb, OnClicked, question, index, Button.Align.left);
-			this.sb = sb;
-		}
+        public PickerButton(Vector2 pos, Vector2 size, SpriteBatch sb, string text, string question, string startsWith, int index, params string[] answers) : 
+            this(sb, text, question, startsWith, answers)
+        {
+            TabIndex = index;
+            Position = pos;
+            Size = size;
+        }
 
-		private void OnClicked(object sender)
-		{
-			++selected;
-			if (selected >= answers.Length)
-				selected = 0;
-		}
+        public PickerButton(SpriteBatch sb, string text, string question, string startsWith, params string[] answers) : base(sb)
+        {
+            for (int i = 0; i < answers.Length; ++i) selected = answers[i] == startsWith ? i : selected;
+            this.question = question;
+            this.answers = answers;
+            this.text = text;
+            this.sb = sb;
 
-		public string GetAnswer() => answers[selected];
+            OnReleased += (_,_) => OnClicked();
+        }
 
-		public override void Draw(GameTime gt)
-		{
-			button.Draw(gt);
-		}
+        private void OnClicked()
+        {
+            ++selected;
+            if (selected >= answers.Length)
+                selected = 0;
+        }
 
-		public override void Update(GameTime gt)
-		{
-			button.text = $"{text}:  {answers[selected]}";
-			button.Update(gt);
-		}
-	}
+        public string GetAnswer() => answers[selected];
+
+        public override void Update(GameTime gt)
+        {
+            Text = $"{text}:  {answers[selected]}";
+
+            base.Update(gt);
+        }
+    }
 }
