@@ -3,12 +3,35 @@ using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Media;
+using System.Collections.Generic;
+using System.Linq;
+using System;
 
 namespace Seihou
 {
-
-    partial class MainState : State
+    internal class MainState : State
     {
+        private enum Status
+        {
+            none = -1,
+            pause = 0,
+            death = 1,
+            gameEnd = 2,
+        }
+
+        private const int buttonSpacing = 30;
+        private const float keyboardWaitTime = 1f;
+        private float keyboardWaitTimer = 0;
+
+        private readonly List<FormHost> menus = new()
+        {
+            new FormHost(),
+            new FormHost(),
+            new FormHost(),
+        };
+
+        private Status currentStatus = Status.none;
+
         //Variables
         private readonly EntityManager _entityManager = new();
         private readonly EntityFactory _entityFactory;
@@ -17,10 +40,6 @@ namespace Seihou
         //SpriteFont font1;
         readonly Player player;
         KeyboardState oldKeyState;
-
-        bool pause = false;
-        bool death = false;
-        bool gameEnd = false;
 
         //Constructor
         public MainState(StateManager sm, ContentManager cm, SpriteBatch sb, GraphicsDeviceManager gdm) : base(sm, cm, sb, gdm)
@@ -55,9 +74,7 @@ namespace Seihou
             _entityManager.Draw(gt);
             UI.Draw(gt, sb, sm, _entityManager, _levelManager);
 
-            if (pause) DrawPauseMenu(gt);
-            if (death) DrawDeathMenu(gt);
-            if (gameEnd) DrawGameEndMenu(gt);
+            DrawMenu(gt);
         }
 
         public override void Update(GameTime gt)
@@ -65,41 +82,28 @@ namespace Seihou
             KeyboardState currentKeyState = Keyboard.GetState();
 
             if ((currentKeyState.IsKeyDown(Global.PauseKey1) && oldKeyState.IsKeyUp(Global.PauseKey1)) || (oldKeyState.IsKeyUp(Global.PauseKey2) && currentKeyState.IsKeyDown(Global.PauseKey2)))
-                pause = !pause;
+            {
+                if (currentStatus == Status.none)
+                    SetStatus(Status.pause);
+                else if(currentStatus == Status.pause)
+                    SetStatus(Status.none);
+            }
 
             oldKeyState = currentKeyState;
 
-            if (!(pause || death))
+            if (currentStatus == Status.none)
             {
                 _levelManager.Update(gt);
                 _entityManager.Update(gt);
 				CloudManager.Instance.Update(gt);
             }
 
-            if (pause) UpdatePauseMenu(gt);
-            if (death) UpdateDeathMenu(gt);
-            if (gameEnd) UpdateGameEndMenu(gt);
-
-			Cursor.Moved();
-			Button.ButtonKeyControl(gt);
+            UpdateMenu(gt);
 		}
 
-        public void OnPlayerDeath()
-        {
-            death = true;
-            if (Keyboard.GetState().IsKeyDown(Settings.GetKey("shootKey")))
-            {
-                Button.AwaitFireKeyUp = true;
-            };
-        }
-        public void OnGameOver()
-        {
-            gameEnd = true;
-            if (Keyboard.GetState().IsKeyDown(Settings.GetKey("shootKey")))
-            {
-                Button.AwaitFireKeyUp = true;
-            };
-        }
+        public void OnPlayerDeath() => SetStatus(Status.death);
+
+        public void OnGameOver() => SetStatus(Status.gameEnd);
 
         //When the state starts
         public override void OnStart()
@@ -116,5 +120,77 @@ namespace Seihou
             MediaPlayer.Stop();
             Debugging.Write(this,"Rip");
         }
+
+        //- Menu logic -
+
+        private FormHost GetHost(Status menu)
+        {
+            if (menu == Status.none)
+                return null;
+
+            return menus[(int)menu];
+        }
+
+        private void UpdateMenu(GameTime gt)
+        {
+            var host = GetHost(currentStatus);
+            if (host is not null)
+            {
+                host.Update(gt);
+                host.BlockKeyboardUserInput = keyboardWaitTimer > 0;
+                keyboardWaitTimer -= gt.Time();
+            }
+        }
+
+        private void DrawMenu(GameTime gt)
+        {
+            if (currentStatus != Status.none)
+                MonoGame.Primitives2D.FillRectangle(sb, new Vector2(0, 0), new Vector2(Global.screenWidth, Global.screenHeight), new Color(Color.Black, 0.6f), 0);
+
+            GetHost(currentStatus)?.Draw(gt);
+        }
+
+        private void BuildMenus()
+        {
+            int width = 500;
+
+            GetHost(Status.death).AddControl(new Button(sb, OnClickedContinue) { Text = "Continue", TabIndex = 0, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y - buttonSpacing * 2) });
+            GetHost(Status.death).AddControl(new Button(sb, OnClickedMenuScore) { Text = "Menu/Save score", TabIndex = 1, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y) });
+            GetHost(Status.death).AddControl(new Button(sb, OnClickedExit) { Text = "Exit", TabIndex = 2, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y + buttonSpacing * 2) });
+
+            GetHost(Status.pause).AddControl(new Button(sb, OnClickedResume) { Text = "Resume", TabIndex = 0, Position = new Vector2(Global.Center.X -width/2, Global.Center.Y - buttonSpacing * 2) });
+            GetHost(Status.pause).AddControl(new Button(sb, OnClickedMenuScore) { Text = "Menu/Save score", TabIndex = 1, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y) });
+            GetHost(Status.pause).AddControl(new Button(sb, OnClickedExit) { Text = "Exit", TabIndex = 2, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y + buttonSpacing * 2) });
+
+            GetHost(Status.gameEnd).AddControl(new Button(sb, OnClickedMenuScore) { Text = "Menu/Save score", TabIndex = 0, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y) });
+            GetHost(Status.gameEnd).AddControl(new Button(sb, OnClickedExit) { Text = "Exit", TabIndex = 1, Position = new Vector2(Global.Center.X - width / 2, Global.Center.Y + buttonSpacing * 2) });
+
+            foreach (var control in menus.SelectMany(m => m.Controls))
+            {
+                var button = control as Button;
+                button.Size = new Vector2(width, buttonSpacing);
+                button.Align = TextAlign.Center;
+            }
+        }
+
+        private void SetStatus(Status status)
+        {
+            currentStatus = status;
+
+            GetHost(currentStatus)?.ResetTabIndex();
+            keyboardWaitTimer = keyboardWaitTime;
+        }
+
+        private void OnClickedContinue()
+        {
+            Global.player.Continue();
+            currentStatus = Status.none;
+        }
+
+        private void OnClickedResume() => currentStatus = Status.none;
+
+        private void OnClickedExit() => sm.abort = true;
+
+        private void OnClickedMenuScore() => sm.ChangeState(new QuestionState(sm, cm, sb, gdm, Math.Round(Global.player.score), Settings.GetDifficulty()));
     }
 }
